Query the jsoncons C++ library

version() reports the version of the C++ jsoncons library in use.

jsonpath() executes a query against a JSON string using the 'jsonpath' specification

jmespath() executes a query against a JSON string using the 'jmespath' specification.

jsonpointer() extracts an element from a JSON string using the 'JSON pointer' specification.

jsonpivot() transforms a JSON array-of-objects to an object-of-arrays; this can be useful when forming a column-based tibble from row-oriented JSON.

as_r() transforms a JSON string to an R object.

Usage

version()

jsonpath(data, path, object_names = "asis", as = "string", ...)

jmespath(data, path, object_names = "asis", as = "string", ...)

jsonpointer(data, path, object_names = "asis", as = "string", ...)

jsonpivot(data, object_names = "asis", as = "string", ...)

as_r(data, object_names = "asis", ...)

Arguments

data

an R object. If data is a scalar (length 1) character vector, it is treated as a single JSON string. Otherwise, it is parsed to a JSON string using jsonlite::toJSON(). Use I() to treat a scalar character vector as an R object rather than JSON string, e.g., I("A") will be parsed to ["A"] before processing.

path

character(1) jsonpath or jmespath query string.

object_names

character(1) order data object elements "asis" (default) or "sort" before filtering on path.

as

character(1) return type. "string" returns a single JSON string; "R" returns an R object following the rules outlined below.

...

arguments passed to jsonlite::toJSON when data is not a scalar character vector. For example, use auto_unbox = TRUE to automatically 'unbox' vectors of length 1 to JSON scalar values.

Value

version() returns a character(1) major.minor.patch version string .

jsonpath(), jmespath() and jsonpivot() return a character(1) JSON string (as = "string", default) or R

object (as = "R") representing the result of the query.

as_r() returns an R object.

Details

jsonpivot() transforms an 'array-of-objects' (typical when the JSON is a row-oriented representation of a table) to an 'object-of-arrays'. A simple example transforms an array of two objects each with three fields '[{"a": 1, "b": 2, "c": 3}, {"a": 4, "b": 5, "c": 6}]' to an object with with three fields, each a vector of length 2 '{"a": [1, 4], "b": [2, 5], "c": [3, 6]}'. The object-of-arrays representation corresponds closely to an R data.frame or tibble, as illustrated in the examples.

The as = "R" argument to jsonpath(), jmespath() and jsonpivot(), and the as_r() function transform a JSON string representation to an R object. Main rules are:

JSON arrays of a single type (boolean, integer, double, string) are transformed to R vectors of the same length and corresponding type. A JSON scalar and a JSON vector of length 1 are represented in the same way in R.

If a JSON 64-bit integer array contains a value larger than R's 32-bit integer representation, the array is transformed to an R numeric vector. NOTE that this results in loss of precision for 64-bit integer values greater than 2^53.

JSON arrays mixing integer and double values are transformed to R numeric vectors.

JSON objects are transformed to R named lists.

The vignette reiterates this information and provides additional details.

Examples

version()
#> [1] "0.173.2"

json <- '{
  "locations": [
    {"name": "Seattle", "state": "WA"},
    {"name": "New York", "state": "NY"},
    {"name": "Bellevue", "state": "WA"},
    {"name": "Olympia", "state": "WA"}
  ]
 }'

## return a JSON string
jsonpath(json, "$..name") |>
    cat("\n")
#> ["Seattle","New York","Bellevue","Olympia"] 

## return an R object
jsonpath(json, "$..name", as = "R")
#> [1] "Seattle"  "New York" "Bellevue" "Olympia" 

## create a list with state and name as scalar vectors
lst <- jsonlite::fromJSON(json, simplifyVector = FALSE)

## objects other than scalar character vectors are automatically
## coerced to JSON; use `auto_unbox = TRUE` to represent R scalar
## vectors in the object as JSON scalar vectors
jsonpath(lst, "$..name", auto_unbox = TRUE) |>
    cat("\n")
#> ["Seattle","New York","Bellevue","Olympia"] 

## a scalar character vector like "Seattle" is not valid JSON...
try(jsonpath("Seattle", "$"))
#> Error : JSON syntax_error at line 1 and column 1
## ...but a double-quoted string is
jsonpath('"Seattle"', "$")
#> [1] "[\"Seattle\"]"

## use I("Seattle") to coerce to a JSON object ["Seattle"]
jsonpath(I("Seattle"), "$[0]")      |> cat("\n")
#> ["Seattle"] 

## different ordering of object names -- 'asis' (default) or 'sort'
json_obj <- '{"b": "1", "a": "2"}'
jsonpath(json_obj, "$")             |> cat("\n")
#> [{"b":"1","a":"2"}] 
jsonpath(json_obj, "$.*")           |> cat("\n")
#> ["1","2"] 
jsonpath(json_obj, "$", "sort")   |> cat("\n")
#> [{"a":"2","b":"1"}] 
jsonpath(json_obj, "$.*", "sort") |> cat("\n")
#> ["2","1"] 

path <- "locations[?state == 'WA'].name | sort(@)"
jmespath(json, path) |>
    cat("\n")
#> ["Bellevue","Olympia","Seattle"] 

## original filter always fails, e.g., '["WA"] != 'WA'
jmespath(lst, path)  # empty result set, '[]'
#> [1] "[]"

## filter with unboxed state, and return unboxed name
jmespath(lst, "locations[?state[0] == 'WA'].name[0] | sort(@)") |>
    cat("\n")
#> ["Bellevue","Olympia","Seattle"] 

## automatically unbox scalar values when creating the JSON string
jmespath(lst, path, auto_unbox = TRUE) |>
    cat("\n")
#> ["Bellevue","Olympia","Seattle"] 

## jsonpointer 0-based arrays
jsonpointer(json, "/locations/0/name")
#> [1] "Seattle"

## document root "", sort selected element keys
jsonpointer('{"b": 0, "a": 1}', "", "sort", as = "R") |>
    str()
#> List of 2
#>  $ a: int 1
#>  $ b: int 0

## 'Key not found' -- path '/' is searches for a 0-length key
try(jsonpointer('{"b": 0, "a": 1}', "/"))
#> Error : Key not found

json |>
    ## 'locations' is a array of objects with 'name' and 'state' scalars...
    jmespath("locations") |>
    ## ...pivot to a single object with 'name' and 'state' vectors...
    jsonpivot(as = "R") |>
    ## ... easily coerced to a data.frame or dplyr::tibble
    as.data.frame()
#>       name state
#> 1  Seattle    WA
#> 2 New York    NY
#> 3 Bellevue    WA
#> 4  Olympia    WA

## as_r()
as_r('[1, 2, 3]')       # JSON integer array -> R integer vector
#> [1] 1 2 3
as_r('[1, 2.0, 3]')     # JSON intger and double array -> R numeric vector
#> [1] 1 2 3
as_r('[1, 2.0, "3"]')   # JSON mixed array -> R list
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] "3"
#> 
as_r('[1, 2147483648]') # JSON integer > R integer max -> R numeric vector
#> [1]          1 2147483648

json = '{"b": 1, "a": ["c", "d"], "e": true, "f": [true], "g": {}}'
as_r(json) |> str()     # parsing complex objects
#> List of 5
#>  $ b: int 1
#>  $ a: chr [1:2] "c" "d"
#>  $ e: logi TRUE
#>  $ f: logi TRUE
#>  $ g: Named list()
identical(              # JSON scalar and length 1 array identical in R
    as_r('{"a": 1}'), as_r('{"a": [1]}')
)
#> [1] TRUE