Patch or compute the difference between two JSON documents

j_patch_apply() uses JSON Patch https://jsonpatch.com to transform JSON 'data' according the rules in JSON 'patch'.

j_patch_from() computes a JSON patch describing the difference between to JSON documents.

Usage

j_patch_apply(data, patch, as = "string", ...)

j_patch_from(data_x, data_y, as = "string", ...)

Arguments

data

JSON character vector, file, URL, or an R object to be converted to JSON using jsonline::fromJSON(data, ...).

patch

JSON 'patch' as character vector, file, URL, or R object.

as

character(1) return type; "string" returns a JSON string, "R" returns an R object using the rules in as_r().

...

passed to jsonlite::toJSON when data, patch, data_x, and / or data_y is an R object. Usually, it is appropriate to add the jsonlite::toJSON() argument auto_unbox = TRUE when patch is an R object (because the elements of the patch objects are scalar-valued, not arrays of length 1).

data_x

As for data.

data_y

As for data.

Value

j_patch_apply() returns a JSON string or R object representing 'data' patched according to 'patch'.

j_patch_from() returns a JSON string or R object representing the difference between 'data_x' and 'data_y'.

Details

For j_patch_apply(), 'patch' is a JSON array of objects. Each object describes how the patch is to be applied. Simple examples are available at https://jsonpatch.com, with verbs 'add', 'remove', 'replace', 'copy' and 'test'. The 'path' element of each operation is a JSON pointer; remember that JSON arrays are 0-based.

add -- add elements to an existing document.

{"op": "add", "path": "/biscuits/1", "value": {"name": "Ginger Nut"}}

remove -- remove elements from a document.

{"op": "remove", "path": "/biscuits/0"}

replace -- replace one element with another

{
    "op": "replace", "path": "/biscuits/0/name",
    "value": "Chocolate Digestive"
}

copy -- copy a path to another location.

{"op": "copy", "from": "/biscuits/0", "path": "/best_biscuit"}

move -- move a path to another location.

{"op": "move", "from": "/biscuits", "path": "/cookies"}

test -- test for the existence of a path; if the path does not exist, do not apply any of the patch.

{"op": "test", "path": "/best_biscuit/name", "value": "Choco Leibniz"}

The examples below illustrate a patch with one (a JSON array with a single object) or several (a JSON array with several arguments) operations. j_patch_apply() fits naturally into a pipeline composed with |> to transform JSON between representations.

Examples

data_file <- system.file(package = "rjsoncons", "extdata", "patch_data.json")

## add a biscuit
patch <- '[
    {"op": "add", "path": "/biscuits/1", "value": {"name": "Ginger Nut"}}
]'
j_patch_apply(data_file, patch, as = "R") |> str()
#> List of 1
#>  $ biscuits:List of 3
#>   ..$ :List of 1
#>   .. ..$ name: chr "Digestive"
#>   ..$ :List of 1
#>   .. ..$ name: chr "Ginger Nut"
#>   ..$ :List of 1
#>   .. ..$ name: chr "Choco Leibniz"

## add a biscuit and choose a favorite
patch <- '[
    {"op": "add", "path": "/biscuits/1", "value": {"name": "Ginger Nut"}},
    {"op": "copy", "from": "/biscuits/2", "path": "/best_biscuit"}
]'
biscuits <- j_patch_apply(data_file, patch)
as_r(biscuits) |> str()
#> List of 2
#>  $ biscuits    :List of 3
#>   ..$ :List of 1
#>   .. ..$ name: chr "Digestive"
#>   ..$ :List of 1
#>   .. ..$ name: chr "Ginger Nut"
#>   ..$ :List of 1
#>   .. ..$ name: chr "Choco Leibniz"
#>  $ best_biscuit:List of 1
#>   ..$ name: chr "Choco Leibniz"

j_patch_from(biscuits, data_file, as = "R") |> str()
#> List of 3
#>  $ :List of 3
#>   ..$ op   : chr "replace"
#>   ..$ path : chr "/biscuits/1/name"
#>   ..$ value: chr "Choco Leibniz"
#>  $ :List of 2
#>   ..$ op  : chr "remove"
#>   ..$ path: chr "/biscuits/2"
#>  $ :List of 2
#>   ..$ op  : chr "remove"
#>   ..$ path: chr "/best_biscuit"